Implemented canvas caching
Every operation caches its output on render and sets it dirty state to false. Whenever an option is changed, it's dirty again and will re-render next time. If it's not dirty, it will just put the cached image data on the canvas.

When rendering, the UI canvas finds the first dirty operation in the stack and sets all following operations to dirty as well so that they don't overwrite the output of the operation that has been changed.